Cast macro variables

+++ Shared/Inc/Libs/_simcom.h
@@ -20,10 +20,10 @@
 #define SIMCOM_RSP_READY                "RDY"
 #define SIMCOM_RSP_IPD                  "+IPD,"
 #define SIMCOM_CMD_BOOT                 "AT\r"
-#define SIMCOM_DEBUG                    1
-#define SIMCOM_MAX_UPLOAD_RETRY         3
+#define SIMCOM_DEBUG                    (uint8_t) 1
+#define SIMCOM_MAX_UPLOAD_RETRY         (uint8_t) 3
 
-#define MAX_ENUM_SIZE                   0xFFFFFFFF
+#define MAX_ENUM_SIZE                   (uint32_t) 0xFFFFFFFF
 
 /* Exported enum -------------------------------------------------------------*/
 typedef enum {
